Update to enable tracking id from order to trade
This is a simple change to backtesting.py to enable a user to track which trade came from which order (and a lot of other potential benefits) as outlined in issue #197.

It simply allows the assignment of an id when creating an order at with `Strategy.buy()` and `Strategy.sell()`.  The id is type agnostic, so the user can make it a string, int, float or whatever may be necessary.

I could not think of a better way to solve this problem and I will be able to test that this works in the next day or so.  I hope people think this is a good idea :worried: :sweat_smile: