DOC: Add return type annotations for buy and sell methods (#975)
* ENH: Explicitly import annotations everywhere

Type annotations are now explicitly imported
from `__future__` to allow forward references
and postponed evaluation (faster import time).
See PEP 563 for details.

* ENH: add return type annotations for `buy`/`sell`

The return type annotations are now added for `buy`
and `sell` methods. The documentation is updated to
mention that the `Order` is returned. Now it should
be crystal clear how to cancel a non-executed order.

This should address #957.